requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiohttp>=3.9.0
tenacity>=8.2.3
structlog>=23.2.0
pyyaml>=6.0.1
schedule>=1.2.0
rich>=13.0.0